Refactor BarChart Component and Move to Common Directory
Related to #1464
