+++ src/app/affiliate/[projectId]/page.tsx
@@ -11,8 +11,7 @@ import { toast } from "react-toastify";
 
 // Components
 import { ProjectHeader, ConversionPointsTable } from "../../components/project";
-import { StatisticCard } from "../../components/common";
-import { BarChart } from "../../components/dashboard";
+import { StatisticCard, BarChart } from "../../components/common";
 
 // Types
 import { ProjectData, ReferralData, ConversionLog, ClickData, Tier } from "../../types";
