+++ src/app/components/common/index.ts
@@ -1,5 +1,7 @@
 import { StatisticCard } from "./StatisticCard";
+import { BarChart } from "./BarChart";
 
 export {
   StatisticCard,
+  BarChart,
 }
