+++ src/app/components/dashboard/index.ts
@@ -1,9 +1,7 @@
 import { NavBar } from "./NavBar";
-import { BarChart } from "./BarChart";
 import { AffiliatesList } from "./AffiliatesList";
 
 export {
   NavBar,
-  BarChart,
   AffiliatesList
 }
